Logistic regression via torch

lantern_logistic_reg() uses regularized estimation to fit models with categorical outcomes with two levels.

Details

For this engine, there is a single mode: classification

Tuning Parameters

This model has 1 tuning parameters listed as a main argument:

penalty: Amount of Regularization (type: double, default: 0.001)

For penalty, the amount of regularization is only L2 penalty (i.e., ridge or weight decay).

There are other parameters that can be set or tuned via parsnip::set_engine():

epochs: The (maximum) number of training iterations (default: 20L)

learn_rate: The rate that the optimization procedure chases the direction of steepest decent (default: 1.0, optimizer = 'SGD' only)

momentum: The amount to adjust the gradient direction/length with the value from a previous iteration. (default: 0.0, optimizer = 'SGD' only)

batch_size: (default: )

stop_iter: The number of iterations with no improvement before stopping. (default: 5L)

class_weights: Numeric class weights. The value for this argument could be:

A named numeric vector (in any order) where the names are the outcome factor levels.

An unnamed numeric vector assumed to be in the same order as the outcome factor levels.

A single numeric value for the least frequent class in the training data and all other classes receive a weight of one.

Other relevant arguments that are not tuning parameters:

validation: The proportion of the data randomly assigned to a validation set. (default: 0.1)

optimizer: The method used in the optimization procedure. Possible choices are 'LBFGS' and 'SGD'. (default: 'LBFGS')

Translation from parsnip to the original package

logistic_reg(penalty = double(1)) %>% 
  set_engine("torch") %>% 
  translate()

## Logistic Regression Model Specification (classification)
## 
## Main Arguments:
##   penalty = double(1)
## 
## Computational engine: torch 
## 
## Model fit template:
## lantern::lantern_logistic_reg(x = missing_arg(), y = missing_arg(), 
##     decay = double(1))

Preprocessing requirements

Factor/categorical predictors need to be converted to numeric values (e.g., dummy or indicator variables) for this engine. When using the formula method via fit.model_spec(), parsnip will convert factor columns to indicators.

Predictors should have the same scale. One way to achieve this is to center and scale each so that each predictor has mean zero and a variance of one.

References

Hoerl, A., & Kennard, R. (2000). Ridge Regression: Biased Estimation for Nonorthogonal Problems. Technometrics, 42(1), 80-86.