Fit a linear regression using torch

lantern_linear_reg() fits a model.

lantern_linear_reg(x, ...)

# S3 method for default
lantern_linear_reg(x, ...)

# S3 method for data.frame
lantern_linear_reg(
  x,
  y,
  epochs = 100L,
  learn_rate = 0.01,
  conv_crit = 0,
  verbose = FALSE,
  ...
)

# S3 method for matrix
lantern_linear_reg(
  x,
  y,
  epochs = 100L,
  learn_rate = 0.01,
  conv_crit = 0,
  verbose = FALSE,
  ...
)

# S3 method for formula
lantern_linear_reg(
  formula,
  data,
  epochs = 100L,
  learn_rate = 0.01,
  conv_crit = 0,
  verbose = FALSE,
  ...
)

# S3 method for recipe
lantern_linear_reg(
  x,
  data,
  epochs = 100L,
  learn_rate = 0.01,
  conv_crit = 0,
  verbose = FALSE,
  ...
)

Arguments

x
Depending on the context:

A data frame of predictors.

A matrix of predictors.

A recipe specifying a set of preprocessing steps created from recipes::recipe().

The predictor data should be standardized (e.g. centered or scaled).

...
Not currently used, but required for extensibility.

y
When x is a data frame or matrix, y is the outcome specified as:

A data frame with 1 numeric column.

A matrix with 1 numeric column.

A numeric vector.

epochs
An integer for the number of epochs of training.

learn_rate
A positive number (usually less than 0.1).

conv_crit
A non-negative number for convergence.

verbose
A logical that prints out the iteration history.

formula
A formula specifying the outcome terms on the left-hand side, and the predictor terms on the right-hand side.

data
When a recipe or formula is used, data is specified as:

A data frame containing both the predictors and the outcome.

Value

A lantern_linear_reg object.

Examples

predictors <- mtcars[, -1]
outcome <- mtcars[, 1]

# XY interface
mod <- lantern_linear_reg(predictors, outcome)

# Formula interface
mod2 <- lantern_linear_reg(mpg ~ ., mtcars)

# Recipes interface
library(recipes)

#> Loading required package: dplyr

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

#> 
#> Attaching package: ‘recipes’

#> The following object is masked from ‘package:stats’:
#> 
#>     step

rec <- recipe(mpg ~ ., mtcars)
rec <- step_log(rec, disp)
mod3 <- lantern_linear_reg(rec, mtcars)