Fit a single layer neural network using torch

lantern_mlp() fits a model.

lantern_mlp() fits a model.

lantern_mlp(x, ...)

# S3 method for default
lantern_mlp(x, ...)

# S3 method for data.frame
lantern_mlp(
  x,
  y,
  epochs = 100L,
  hidden_units = 3L,
  activation = "relu",
  penalty = 0,
  dropout = 0,
  validation = 0.1,
  learn_rate = 0.01,
  momentum = 0,
  batch_size = NULL,
  conv_crit = -Inf,
  verbose = FALSE,
  ...
)

# S3 method for matrix
lantern_mlp(
  x,
  y,
  epochs = 100L,
  hidden_units = 3L,
  activation = "relu",
  penalty = 0,
  dropout = 0,
  validation = 0.1,
  learn_rate = 0.01,
  momentum = 0,
  batch_size = NULL,
  conv_crit = -Inf,
  verbose = FALSE,
  ...
)

# S3 method for formula
lantern_mlp(
  formula,
  data,
  epochs = 100L,
  hidden_units = 3L,
  activation = "relu",
  penalty = 0,
  dropout = 0,
  validation = 0.1,
  learn_rate = 0.01,
  momentum = 0,
  batch_size = NULL,
  conv_crit = -Inf,
  verbose = FALSE,
  ...
)

# S3 method for recipe
lantern_mlp(
  x,
  data,
  epochs = 100L,
  hidden_units = 3L,
  activation = "relu",
  penalty = 0,
  dropout = 0,
  validation = 0.1,
  learn_rate = 0.01,
  momentum = 0,
  batch_size = NULL,
  conv_crit = -Inf,
  verbose = FALSE,
  ...
)

lantern_mlp(x, ...)

# S3 method for default
lantern_mlp(x, ...)

# S3 method for data.frame
lantern_mlp(
  x,
  y,
  epochs = 100L,
  hidden_units = 3L,
  activation = "relu",
  penalty = 0,
  dropout = 0,
  validation = 0.1,
  learn_rate = 0.01,
  momentum = 0,
  batch_size = NULL,
  conv_crit = -Inf,
  verbose = FALSE,
  ...
)

# S3 method for matrix
lantern_mlp(
  x,
  y,
  epochs = 100L,
  hidden_units = 3L,
  activation = "relu",
  penalty = 0,
  dropout = 0,
  validation = 0.1,
  learn_rate = 0.01,
  momentum = 0,
  batch_size = NULL,
  conv_crit = -Inf,
  verbose = FALSE,
  ...
)

# S3 method for formula
lantern_mlp(
  formula,
  data,
  epochs = 100L,
  hidden_units = 3L,
  activation = "relu",
  penalty = 0,
  dropout = 0,
  validation = 0.1,
  learn_rate = 0.01,
  momentum = 0,
  batch_size = NULL,
  conv_crit = -Inf,
  verbose = FALSE,
  ...
)

# S3 method for recipe
lantern_mlp(
  x,
  data,
  epochs = 100L,
  hidden_units = 3L,
  activation = "relu",
  penalty = 0,
  dropout = 0,
  validation = 0.1,
  learn_rate = 0.01,
  momentum = 0,
  batch_size = NULL,
  conv_crit = -Inf,
  verbose = FALSE,
  ...
)

Arguments

x
Depending on the context:

A data frame of predictors.

A matrix of predictors.

A recipe specifying a set of preprocessing steps created from recipes::recipe().

The predictor data should be standardized (e.g. centered or scaled).

...
Not currently used, but required for extensibility.

y
When x is a data frame or matrix, y is the outcome specified as:

A data frame with 1 numeric column.

A matrix with 1 numeric column.

A numeric vector.

epochs
An integer for the number of epochs of training.

hidden_units
An integer for the number of hidden units.

activation
A string for the activation function. Possible values are "relu", "elu", "tanh", and "linear".

penalty
The amount of weight decay (i.e., L2 regularization).

dropout
The proportion of parameters set to zero.

validation
The proportion of the data randomly assigned to a validation set.

learn_rate
A positive number (usually less than 0.1).

momentum
A positive number on [0, 1] for the momentum parameter in gradient decent.

batch_size
An integer for the number of training set points in each batch.

conv_crit
A non-negative number for convergence.

verbose
A logical that prints out the iteration history.

formula
A formula specifying the outcome terms on the left-hand side, and the predictor terms on the right-hand side.

data
When a recipe or formula is used, data is specified as:

A data frame containing both the predictors and the outcome.

Value

A lantern_mlp object with elements:

models: a list object of serialized models for each epoch.

loss: A vector of loss values (MSE for regression, negative log- likelihood for classification) at each epoch.

dim: A list of data dimensions.

parameters: A list of some tuning parameter values.

blueprint: The hardhat blueprint data.

A lantern_mlp object with elements:

models: a list object of serialized models for each epoch.

loss: A vector of loss values (MSE for regression, negative log- likelihood for classification) at each epoch.

dim: A list of data dimensions.

y_stats: A list of summary statistics for numeric outcomes.

parameters: A list of some tuning parameter values.

blueprint: The hardhat blueprint data.

Details

This function fits single layer, feed-forward neural network models for regression (when the outcome is a number) or classification (a factor). For regression, the mean squared error is optimized and cross-entropy is the loss function for classification.

The predictors data should all be numeric and encoded in the same units (e.g. standardized to the same range or distribution). If there are factor predictors, use a recipe or formula to create indicator variables (or some other method) to make them numeric.

If conv_crit is used, it stops training when the difference in the loss function is below conv_crit or if it gets worse. The default trains the model over the specified number of epochs.

This function fits single layer, feed-forward neural network models for regression (when the outcome is a number) or classification (a factor). For regression, the mean squared error is optimized and cross-entropy is the loss function for classification.

The predictors data should all be numeric and encoded in the same units (e.g. standardized to the same range or distribution). If there are factor predictors, use a recipe or formula to create indicator variables (or some other method) to make them numeric.

When the outcome is a number, the function internally standardizes the outcome data to have mean zero and a standard deviation of one. The prediction function creates predictions on the original scale.

If conv_crit is used, it stops training when the difference in the loss function is below conv_crit or if it gets worse. The default trains the model over the specified number of epochs.

Examples

# \donttest{
if (torch::torch_is_installed()) {

 ## -----------------------------------------------------------------------------
 # regression examples (increase # epochs to get better results)

 data(ames, package = "modeldata")

 ames$Sale_Price <- log10(ames$Sale_Price)

 set.seed(122)
 in_train <- sample(1:nrow(ames), 2000)
 ames_train <- ames[ in_train,]
 ames_test  <- ames[-in_train,]

 # Using matrices
 set.seed(1)
 lantern_mlp(x = as.matrix(ames_train[, c("Longitude", "Latitude")]),
           y = ames_train$Sale_Price,
           penalty = 0.10, epochs = 20, batch_size = 32)

 # Using recipe
 library(recipes)

 ames_rec <-
  recipe(Sale_Price ~ Longitude + Latitude + Alley, data = ames_train) %>%
  step_dummy(Alley) %>%
  step_normalize(all_predictors())

 set.seed(1)
 fit <- lantern_mlp(ames_rec, data = ames_train,
                  dropout = 0.25, epochs = 20, batch_size = 32)
 fit

 autoplot(fit)

 predict(fit, ames_test)
 }

#> # A tibble: 930 x 1
#>    .pred
#>    <dbl>
#>  1  5.22
#>  2  5.22
#>  3  5.33
#>  4  5.32
#>  5  5.31
#>  6  5.29
#>  7  5.29
#>  8  5.26
#>  9  5.26
#> 10  5.23
#> # … with 920 more rows

# }
# \donttest{
if (torch::torch_is_installed()) {

 ## -----------------------------------------------------------------------------
 # regression examples (increase # epochs to get better results)

 data(ames, package = "modeldata")

 ames$Sale_Price <- log10(ames$Sale_Price)

 set.seed(122)
 in_train <- sample(1:nrow(ames), 2000)
 ames_train <- ames[ in_train,]
 ames_test  <- ames[-in_train,]

 # Using matrices
 set.seed(1)
 lantern_mlp(x = as.matrix(ames_train[, c("Longitude", "Latitude")]),
           y = ames_train$Sale_Price,
           penalty = 0.10, epochs = 20, batch_size = 32)

 # Using recipe
 library(recipes)

ames_rec <-
 recipe(Sale_Price ~ Bldg_Type + Neighborhood + Year_Built + Gr_Liv_Area +
         Full_Bath + Year_Sold + Lot_Area + Central_Air + Longitude + Latitude,
        data = ames_train) %>%
 # Transform some highly skewed predictors
 step_BoxCox(Lot_Area, Gr_Liv_Area) %>%
 # Lump some rarely occuring categories into "other"
 step_other(Neighborhood, threshold = 0.05)  %>%
 # Encode categorical predictors as binary.
 step_dummy(all_nominal(), one_hot = TRUE) %>%
 # Add an interaction effect:
 step_interact(~ starts_with("Central_Air"):Year_Built) %>%
 step_zv(all_predictors()) %>%
 step_normalize(all_predictors())

set.seed(2)
fit <- lantern_mlp(ames_rec, data = ames_train, hidden_units = 20,
                 dropout = 0.05, epochs = 20, batch_size = 32)
fit

autoplot(fit)

library(ggplot2)

predict(fit, ames_test) %>%
 bind_cols(ames_test) %>%
 ggplot(aes(x = .pred, y = Sale_Price)) +
 geom_abline(col = "green") +
 geom_point(alpha = .3) +
 lims(x = c(4, 6), y = c(4, 6)) +
 coord_fixed(ratio = 1)

library(yardstick)
predict(fit, ames_test) %>%
 bind_cols(ames_test) %>%
 rmse(Sale_Price, .pred)
 }

#> For binary classification, the first factor level is assumed to be the event.
#> Use the argument `event_level = "second"` to alter this as needed.

#> # A tibble: 1 x 3
#>   .metric .estimator .estimate
#>   <chr>   <chr>          <dbl>
#> 1 rmse    standard     0.00699

# }